NewsTeam1/requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arviz @ file:///tmp/build/80754af9/arviz_1614019183254/work
asgiref==3.5.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
cftime @ file:///C:/ci/cftime_1638358679566/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1655412516417/work
convertdate @ file:///home/conda/feedstock_root/build_artifacts/convertdate_1642883757836/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
Cython @ file:///D:/bld/cython_1659102065258/work
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-bootstrap5==22.2
django-environ==0.9.0
djangorestframework==3.14.0
entrypoints==0.4
ephem @ file:///C:/ci/ephem_1638960661180/work
fastjsonschema==2.16.2
fbprophet @ file:///D:/bld/fbprophet_1599365723319/work
hijri-converter @ file:///home/conda/feedstock_root/build_artifacts/hijri-converter_1653348285237/work
holidays @ file:///home/conda/feedstock_root/build_artifacts/holidays_1663340966408/work
idna==3.4
importlib-metadata @ file:///D:/bld/importlib-metadata_1653252990345/work
importlib-resources==5.10.0
ipykernel==6.16.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver @ file:///D:/bld/kiwisolver_1657953189205/work
korean-lunar-calendar @ file:///home/conda/feedstock_root/build_artifacts/korean_lunar_calendar_1663341251025/work
LunarCalendar==0.0.9
MarkupSafe==2.1.1
matplotlib @ file:///D:/bld/matplotlib-base_1594091882257/work
matplotlib-inline==0.1.6
mistune==2.0.4
mysqlclient==2.0.1
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
netCDF4 @ file:///C:/ci/netcdf4_1643100796549/work
notebook==6.5.1
notebook-shim==0.1.0
numpy @ file:///D:/bld/numpy_1649806536290/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas @ file:///D:/bld/pandas_1614698123791/work
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.31
psutil==5.9.2
pycparser==2.21
Pygments==2.13.0
PyMeeus @ file:///home/conda/feedstock_root/build_artifacts/pymeeus_1615809745702/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.18.1
pystan==2.19.1.1
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1664798238822/work
pywin32==304
pywinpty==2.0.8
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
scipy @ file:///C:/bld/scipy_1637806996411/work
Send2Trash==1.8.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
terminado==0.16.0
tinycss2==1.1.1
tornado @ file:///D:/bld/tornado_1656937938087/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
traitlets==5.4.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==4.0.3
wincertstore==0.2
xarray @ file:///home/conda/feedstock_root/build_artifacts/xarray_1639125986756/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1665306627898/work
